refactor(markdown): 收敛文件预览渲染边界
