+++ openspec/changes/archive/2026-07-25-separate-markdown-preview-renderer-boundaries/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-07-25
